+++ src/modules/admissions/queries.ts
@@ -469,6 +469,7 @@ export const GET_APPLICATION_FROM_TOKEN = gql`
         internalGroupPosition {
           id
           name
+          admissionMembershipType
         }
       }
       interview {
@@ -489,6 +490,7 @@ export const INTERNAL_GROUP_POSITIONS_AVAILABLE_FOR_APPLICANTS_QUERY = gql`
     internalGroupPositionsAvailableForApplicants {
       id
       name
+      admissionMembershipType
     }
   }
 `